unique_ptr, "constify_ref" in Assembly
Using std::unique_ptr here cleans up the destructor and makes the code
clearer.

The "constify_ref" trick is a much cleaner way of convincing a C++
compiler to give us a T * const & into a const T * const &; it really is
the same address underneath, so we don't need to maintain a copy of
everything, we just need to convince the compiler that it doesn't need
to try to make a temporary copy either.